requirements.txt
# requirements.txt

asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2024.1
drf-nested-routers==0.93.5
django-rest-knox==4.2.0